规范maven项目结构并升级版本号(1.27-SNAPSHOT)
