+++ README
@@ -1,27 +1,27 @@
 Copyright (C) 2008 Happy Fish / YuQing
 
-FastDFS Java Client API may be copied only under the terms of 
+FastDFS Java Client API may be copied only under the terms of
 the BSD license.
 Please visit the FastDFS Home Page for more detail.
 English language: http://english.csource.org/
 Chinese language: http://www.csource.org/
 
 
-The jar file is compiled by JDK1.5, you can download the last version 
+The jar file is compiled by JDK1.5, you can download the last version
 from google code: http://code.google.com/p/fastdfs/downloads/list
 
 run the FastDFS Java Client test:
 java -cp <fastdfs_client_jar_filename> org.csource.fastdfs.test.TestClient <config_filename> <upload_filename>
 
 eg.:
-java -cp fastdfs_client_v1.25.jar org.csource.fastdfs.test.TestClient fdfs_client.conf c:\windows\system32\notepad.exe
+java -cp <fastdfs_client_jar_filename> org.csource.fastdfs.test.TestClient fdfs_client.conf c:\windows\system32\notepad.exe
 
 or:
-java -cp fastdfs_client_v1.25.jar org.csource.fastdfs.test.TestClient fdfs_client.conf /usr/include/stdlib.h
+java -cp <fastdfs_client_jar_filename> org.csource.fastdfs.test.TestClient fdfs_client.conf /usr/include/stdlib.h
 
 
 run the FastDFS monitor:
 java -cp <fastdfs_client_jar_filename> org.csource.fastdfs.test.Monitor <config_filename>
 
 eg.:
-java -cp fastdfs_client_v1.25.jar org.csource.fastdfs.test.Monitor fdfs_client.conf
+java -cp <fastdfs_client_jar_filename> org.csource.fastdfs.test.Monitor fdfs_client.conf
