+++ src/main/java/org/csource/fastdfs/ClientGlobal.java
