+++ src/main/java/org/csource/fastdfs/ProtoCommon.java
