+++ src/main/java/org/csource/fastdfs/StructBase.java
