+++ src/main/java/org/csource/fastdfs/StructStorageStat.java
